add escape hatch to all terraform elements
